VirtuSig/requirements.txt
Django==3.2.13
django-cors-headers==3.13.0
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
mysqlclient==2.1.0
PyJWT==1.7.1
cryptography==37.0.2
boto3~=1.26.6

